feat: add shouldStop method in the Graph.IteratorVisitor to allow breaks

+++ src/main/java/graph/api/Graph.java
@@ -27,7 +27,7 @@ public abstract class Graph implements GraphBase {
   }
 
   @Override
-  public long getVerticesCount() {
+  public int getVerticesCount() {
     return n;
   }
